Tests: Work around PHP 7.2 stupidity
PHP 7.2 changed various session-related behavior, and part of that was
making ini_set of session settings complain if headers have already been
sent. This makes sense in most cases.

Unfortunately, they applied this even in CLI mode, which doesn't *have*
headers to be sent but still flags "headers sent" when something outputs
to the console. Such as PHPUnit does to print a banner and progress
indicators. Sigh.

Bug: T193957
Change-Id: If698c2c53e0b3d86c3bf8d5a4ca08f84d575375c