Allow newer versions of psr/log
The code of `PhpSessionSerializer` is compatible with all versions of psr/log.
Versions ^2.0 restrict the input type of the methods in the `LoggerInterface` to
`string|\Stringable`, which is fine, because everything that the
`PhpSessionSerializer` passes to those methods is a string. Versions ^3.0 of the
package restrict the return type of the methods in the `LoggerInterface` to
`void`, which is fine, because the `PhpSessionSerializer` does not expect any of
those methods to return anything. Since versions ^1.0 are still allowed this
change is backwards compatible.

Taken from https://github.com/wikimedia/php-session-serializer/pull/1

Adjustments:
 * Altered TestLogger->log() to add an explicit void return type, for 3.0 compat.

Change-Id: I52e02e876e76ca7ce7272039618cb93d4adb37c2